提出課題作成中２_1

+++ app/src/main/java/jp/techacademy/kouhei/konishi/calcapp/SecondActivity.kt
@@ -2,6 +2,7 @@ package jp.techacademy.kouhei.konishi.calcapp
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import kotlinx.android.synthetic.main.activity_second.*
 
 class SecondActivity : AppCompatActivity() {
@@ -10,9 +11,8 @@ class SecondActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_second)
 
-        val value1 = intent.getIntExtra("VALUE1", 0)
-        val value2 = intent.getIntExtra("VALUE2", 0)
-
+        val value1 = intent.getDoubleExtra("VALUE1", 0.0)
+        val value2 = intent.getDoubleExtra("VALUE2", 0.0)
         textView.text = "${value1 + value2}"
     }
 }
